Pull out data transfer.
