+++ src/ansys/fluent/core/session/base_meshing.py
@@ -40,6 +40,7 @@ from typing import TYPE_CHECKING, Any, cast
 
 from ansys.fluent.core._data_model_cache import DataModelCache, NameKey
 from ansys.fluent.core._types import PathType
+from ansys.fluent.core.data_transfer import transfer_case
 from ansys.fluent.core.exceptions import BetaFeaturesNotEnabled
 from ansys.fluent.core.fluent_connection import FluentConnection
 from ansys.fluent.core.module_config import config
@@ -50,7 +51,6 @@ from ansys.fluent.core.session._shared import (
     _make_tui_module,
 )
 from ansys.fluent.core.session.session import BaseSession
-from ansys.fluent.core.utils.data_transfer import transfer_case
 from ansys.fluent.core.utils.fluent_version import (
     get_version_for_file_name,
 )
